adding manual free function for slurm_list_create
This C function requires that it be passed a function that will be used
to free the item memory. previously we were passing null, so nothing was
getting freed in there, but destroy_qos was trying to do the free
itself, and the heap got corrupted. This fix should hopefully stop the
segfaults